python: Move katana.loops operations to katana. Eliminate import cycle.
