+++ docs/index.rst
@@ -3,15 +3,16 @@ Galois Parallel Programming System
 ==================================
 
 .. automodule:: katana
+   :no-index:
    :no-members:
 
 :Version: |release|
 
 .. toctree::
+   katana
    katana.analytics
    katana.atomic
    katana.datastructures
-   katana.loops
    katana.property_graph
    katana.timer
