+++ docs/katana.rst
@@ -2,7 +2,7 @@
 Parallel Loops
 ==============
 
-.. automodule:: katana.loops
+.. automodule:: katana
    :members:
    :undoc-members:
    :special-members: __init__, __iter__, __getitem__, __setitem__, __len__
