+++ python/katana/lonestar/analytics/connected_components.py
@@ -1,9 +1,9 @@
 import numpy as np
 import pyarrow
 
+from katana import do_all, do_all_operator
 from katana.atomic import GAccumulator, GReduceLogicalOr, atomic_min
 from katana.galois import set_active_threads
-from katana.loops import do_all, do_all_operator
 from katana.property_graph import PropertyGraph
 from katana.timer import StatTimer
