+++ python/katana/lonestar/analytics/jaccard.py
@@ -1,8 +1,8 @@
 import numpy as np
 import pyarrow
 
+from katana import do_all, do_all_operator
 from katana.galois import set_active_threads
-from katana.loops import do_all, do_all_operator
 from katana.property_graph import PropertyGraph
 from katana.timer import StatTimer
