+++ python/katana/lonestar/analytics/kcore.py
@@ -1,10 +1,10 @@
 import numpy as np
 import pyarrow
 
+from katana import do_all, do_all_operator, for_each, for_each_operator
 from katana.atomic import GAccumulator, atomic_sub
 from katana.datastructures import AllocationPolicy, InsertBag, NUMAArray
 from katana.galois import set_active_threads
-from katana.loops import do_all, do_all_operator, for_each, for_each_operator
 from katana.property_graph import PropertyGraph
 from katana.timer import StatTimer
