+++ python/katana/lonestar/analytics/pagerank.py
@@ -1,10 +1,10 @@
 import numpy as np
 import pyarrow
 
+from katana import do_all, do_all_operator
 from katana.atomic import GAccumulator, GReduceLogicalOr, GReduceMax, GReduceMin, atomic_add
 from katana.datastructures import AllocationPolicy, NUMAArray
 from katana.galois import set_active_threads
-from katana.loops import do_all, do_all_operator
 from katana.property_graph import PropertyGraph
 from katana.timer import StatTimer
